Modified monthly_liquidate_rebalance_backtest.py to work with the new Backtest API. Included derived Strategy and PositionSizer classes directly in the example file to keep it self-contained. Removed the associated Python files for each of these classes within their own separate directories.
