--- qstrader/strategy/monthly_liquidate_rebalance_strategy.py
@@ -1,57 +1,0 @@
-import calendar
-
-from .base import AbstractStrategy
-
-from ..event import (SignalEvent, EventType)
-
-
-class MonthlyLiquidateRebalanceStrategy(AbstractStrategy):
-    """
-    A generic strategy that allows monthly rebalancing of a
-    set of tickers, via full liquidation and dollar-weighting
-    of new positions.
-
-    Must be used in conjunction with the
-    LiquidateRebalancePositionSizer object to work correctly.
-    """
-    def __init__(self, tickers, events_queue):
-        self.tickers = tickers
-        self.events_queue = events_queue
-        self.tickers_invested = self._create_invested_list()
-
-    def _end_of_month(self, cur_time):
-        """
-        Determine if the current day is at the end of the month.
-        """
-        cur_day = cur_time.day
-        end_day = calendar.monthrange(cur_time.year, cur_time.month)[1]
-        return cur_day == end_day
-
-    def _create_invested_list(self):
-        """
-        Create a dictionary with each ticker as a key, with
-        a boolean value depending upon whether the ticker has
-        been "invested" yet. This is necessary to avoid sending
-        a liquidation signal on the first allocation.
-        """
-        tickers_invested = {ticker: False for ticker in self.tickers}
-        return tickers_invested
-
-    def calculate_signals(self, event):
-        """
-        For a particular received BarEvent, determine whether
-        it is the end of the month (for that bar) and generate
-        a liquidation signal, as well as a purchase signal,
-        for each ticker.
-        """
-        if (
-            event.type in [EventType.BAR, EventType.TICK] and
-            self._end_of_month(event.time)
-        ):
-            ticker = event.ticker
-            if self.tickers_invested[ticker]:
-                liquidate_signal = SignalEvent(ticker, "EXIT")
-                self.events_queue.put(liquidate_signal)
-            long_signal = SignalEvent(ticker, "BOT")
-            self.events_queue.put(long_signal)
-            self.tickers_invested[ticker] = True
